Some refinements of the Power and Int64 fixes

+++ src/miscprocs.c
@@ -583,7 +583,7 @@ int32 TOINT(float64 fltmp) {
 }
 
 int64 TOINT64(float64 fltmp) {
-  if (((uint64)fltmp > MAXINT64VAL) || ((int64)fltmp < MININT64VAL)) error(ERR_RANGE);
+  if ((fltmp > MAXINT64VAL) || (fltmp < MININT64VAL)) error(ERR_RANGE);
   return (int64)fltmp;
 }
